change Ad Duration to double

+++ playkit/src/main/java/com/kaltura/playkit/ads/PKAdInfo.java
@@ -12,6 +12,6 @@ public interface PKAdInfo {
     int getAdWidth();
     int getAdHeight();
     int getAdPodCount();
-    long getAdDuration();
+    double getAdDuration();
     List<Float> getAdCuePoints();
 }
